refactor(protocol-designer, components): delete outdated timeline scrubber ff
