+++ components/src/organisms/CommandText/__tests__/CommandText.test.tsx
@@ -9,9 +9,8 @@ import {
 
 import { i18n } from '../../../i18n'
 import { renderWithProviders } from '../../../testing/utils'
-import { getCommandTextData } from '../../ProtocolTimelineScrubber/utils'
 import { mockRobotSideAnalysis } from '../fixtures'
-import { CommandText } from '../index'
+import { CommandText, getCommandTextData } from '../index'
 
 import type {
   AspirateInPlaceRunTimeCommand,
@@ -33,7 +32,7 @@ import type {
   PrepareToAspirateRunTimeCommand,
   RunTimeCommand,
 } from '@opentrons/shared-data'
-import type { CommandTextData } from '../../ProtocolTimelineScrubber'
+import type { CommandTextData } from '../index'
 
 const mockCommandTextData: CommandTextData = {
   commands: mockRobotSideAnalysis.commands,
