+++ components/src/organisms/CommandText/index.tsx
@@ -14,14 +14,16 @@ import type {
   RunTimeCommand,
 } from '@opentrons/shared-data'
 import type { StyleProps } from '../../primitives'
-import type { CommandTextData } from '../ProtocolTimelineScrubber/types'
 import type {
+  CommandTextData,
   GetTCRunExtendedProfileCommandTextResult,
   GetTCRunProfileCommandTextResult,
   GetTCStartRunExtendedProfileCommandTextResult,
 } from './useCommandTextString'
 
 export * from './useCommandTextString'
+export * from './getCommandData'
+
 interface LegacySTProps {
   as?: ComponentProps<typeof LegacyStyledText>['as']
   modernStyledTextDefaults?: false
