--- components/src/organisms/ProtocolTimelineScrubber/types.ts
@@ -1,6 +1,0 @@
-import type { CompletedProtocolAnalysis } from '@opentrons/shared-data'
-
-export type CommandTextData = Pick<
-  CompletedProtocolAnalysis,
-  'pipettes' | 'labware' | 'modules' | 'liquids' | 'commands'
->
