+++ components/src/organisms/index.ts
@@ -1,6 +1,5 @@
 export * from './DeckLabelSet'
 export * from './Toolbox'
-export * from './ProtocolTimelineScrubber'
 export * from './CommandText'
 export * from './FixtureOption'
 export * from './ProtocolDeck'
