+++ protocol-designer/src/assets/localization/en/feature_flags.json
@@ -27,10 +27,6 @@
     "title": "Enable React Scan",
     "description": "Enable React Scan support for components rendering check"
   },
-  "OT_PD_ENABLE_TIMELINE_SCRUBBER": {
-    "title": "Enable timeline scrubber",
-    "description": "See the protocol timeline visualization in overview"
-  },
   "OT_PD_ENABLE_MULTIPLE_TEMPS_OT2": {
     "title": "Allow two temperature modules on OT-2",
     "description": "This experimental setting may cause collisions, and Opentrons will not be responsible for any damage resulting from its use."
