+++ protocol-designer/src/feature-flags/reducers.ts
@@ -33,8 +33,6 @@ const initialFlags: Flags = {
     _FF_ENV_VARS_.OT_PD_ENABLE_REACT_SCAN === '1' || false,
   OT_PD_ENABLE_MULTIPLE_TEMPS_OT2:
     _FF_ENV_VARS_.OT_PD_ENABLE_MULTIPLE_TEMPS_OT_ === '1' || false,
-  OT_PD_ENABLE_TIMELINE_SCRUBBER:
-    _FF_ENV_VARS_.OT_PD_ENABLE_TIMELINE_SCRUBBER === '1' || false,
   OT_PD_ENABLE_PARTIAL_TIP_SUPPORT:
     _FF_ENV_VARS_.OT_PD_ENABLE_PARTIAL_TIP_SUPPORT === '1' || false,
   OT_PD_ENABLE_STACKING: _FF_ENV_VARS_.OT_PD_ENABLE_STACKING === '1' || false,
