+++ protocol-designer/src/feature-flags/selectors.ts
@@ -38,10 +38,6 @@ export const getEnableMutlipleTempsOT2: Selector<boolean> = createSelector(
   getFeatureFlagData,
   flags => flags.OT_PD_ENABLE_MULTIPLE_TEMPS_OT2 ?? false
 )
-export const getEnableTimelineScrubber: Selector<boolean> = createSelector(
-  getFeatureFlagData,
-  flags => flags.OT_PD_ENABLE_TIMELINE_SCRUBBER ?? false
-)
 export const getEnablePartialTipSupport: Selector<boolean> = createSelector(
   getFeatureFlagData,
   flags => flags.OT_PD_ENABLE_PARTIAL_TIP_SUPPORT ?? false
