+++ protocol-designer/src/feature-flags/types.ts
@@ -33,6 +33,7 @@ export const DEPRECATED_FLAGS = [
   'OT_PD_ALLOW_ALL_TIPRACKS',
   'OT_PD_ENABLE_CAMERA_SUPPORT',
   'OT_PD_ENABLE_JSON_EXPORT',
+  'OT_PD_ENABLE_TIMELINE_SCRUBBER',
 ]
 // union of feature flag string constant IDs
 export type FlagTypes =
@@ -42,7 +43,6 @@ export type FlagTypes =
   | 'OT_PD_ENABLE_HOT_KEYS_DISPLAY'
   | 'OT_PD_ENABLE_REACT_SCAN'
   | 'OT_PD_ENABLE_MULTIPLE_TEMPS_OT2'
-  | 'OT_PD_ENABLE_TIMELINE_SCRUBBER'
   | 'OT_PD_ENABLE_PARTIAL_TIP_SUPPORT'
   | 'OT_PD_ENABLE_STACKING'
   | 'OT_PD_ENABLE_CONCURRENT_MODULE_ACTIONS'
@@ -60,7 +60,6 @@ export const allFlags: FlagTypes[] = [
   'PRERELEASE_MODE',
   'OT_PD_ENABLE_COMMENT',
   'OT_PD_ENABLE_REACT_SCAN',
-  'OT_PD_ENABLE_TIMELINE_SCRUBBER',
   'OT_PD_ENABLE_PARTIAL_TIP_SUPPORT',
   'OT_PD_ENABLE_STACKING',
   'OT_PD_ENABLE_CONCURRENT_MODULE_ACTIONS',
